git-svn-id: https://subversion.assembla.com/svn/mediasl/Projectes%20MPLAB/EOS/trunk@5352 4ad52a4b-eac5-49b1-a5a4-524d969995ed

+++ v2/Source/Services/eosDigInput.cpp
@@ -100,7 +100,8 @@ DigInput::DigInput(
     state = pinGet();
     pattern = state ? 0xFFFFFFFF : 0x00000000;
     
-    service->add(this);
+    if (service != nullptr)
+        service->add(this);
 }
 
 
@@ -109,7 +110,8 @@ DigInput::DigInput(
 ///
 DigInput::~DigInput() {
     
-    service->remove(this);
+    if (service != nullptr)
+        service->remove(this);
     
     if (evChange != nullptr)
         delete evChange;
